+++ src/components/modal/TodoModal/view/ActionButtonView.tsx
@@ -1,8 +1,7 @@
 import React from "react"
-import { StyleSheet, Text, Pressable, ActivityIndicator } from "react-native"
-import theme from "../../../../shared/theme"
 
 import { ModalType } from "../../../../shared/types"
+import Button from "../../../ui/Button"
 
 interface Props {
   isLoading: boolean
@@ -19,47 +18,15 @@ export default function ActionButtonView({
 }: Props) {
   if (modalType === "ADD") {
     return (
-      <Pressable
-        style={styles.button}
-        onPress={handleAddTodo}
-        disabled={isLoading}
-      >
-        {isLoading ? (
-          <ActivityIndicator />
-        ) : (
-          <Text style={styles.textStyle}>추가하기</Text>
-        )}
-      </Pressable>
+      <Button onPress={handleAddTodo} text="추가하기" isLoading={isLoading} />
     )
   }
 
   if (modalType === "EDIT") {
     return (
-      <Pressable
-        style={styles.button}
-        onPress={handleEditTodo}
-        disabled={isLoading}
-      >
-        {isLoading ? (
-          <ActivityIndicator />
-        ) : (
-          <Text style={styles.textStyle}>수정하기</Text>
-        )}
-      </Pressable>
+      <Button onPress={handleEditTodo} text="수정하기" isLoading={isLoading} />
     )
   }
 
   return null
 }
-
-const styles = StyleSheet.create({
-  button: {
-    padding: 10,
-  },
-  textStyle: {
-    color: theme.primary,
-    fontWeight: "bold",
-    textAlign: "center",
-    fontSize: 20,
-  },
-})
